feat: start Main componentization, add basic pages and favicon
