+++ index.html
@@ -2,12 +2,12 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" href="/favicon/favicon-light.ico" media="(prefers-color-scheme: light)" />
+    <link rel="icon" href="/favicon/favicon-dark.ico" media="(prefers-color-scheme: dark)" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite App</title>
+    <title>Dootzy</title>
   </head>
   <body>
-    <div id="app"></div>
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
