+++ src/components/Navbar/Navbar.css
@@ -9,32 +9,37 @@ nav > #logoContainer{
   display: flex;
   align-items: center;
   height: 100%;
-
+  width: 20%;
 }
 
 nav > #logoContainer > img{
-  height: 100%;
+  height: 80%;
   width: auto;
 }
-#menu {
-  padding: 10px;
-}
-
 
 nav > ul {
   display: flex;
   align-items: center;
+  justify-content: end;
   gap: 5rem;
+  width: 60%;
+}
+
+nav li{
+  width: 100%;
+  display: flex;
+  justify-content: end;
 }
 
 nav a {
   color: var(--text);
-  font-size: 1.3rem;
+  font-size: 1.2rem;
   font-weight: 500;
-  transition: all 0.3s ease-in-out;}
+  transition: all 0.3s ease-in-out;
+}
 
 nav a:hover {
-  text-decoration: underline;
+  font-weight: 600;
 }
 
 @media (max-width: 770px) {
